apps/workers/imessage-generator/requirements.txt
Flask~=3.0.3
Pillow~=10.3.0
Flask-WTF~=1.2.1
python-dotenv~=1.0.1
Werkzeug~=3.0.2
WTForms~=3.1.2
elevenlabs~=1.6.0
ffmpeg-python~=0.2.0
pydub~=0.25.1
pilmoji==2.0.4
emoji==2.11.1